feat: Rename export /register to /manager for managed mode in Storybook 10
The export /register stopped working with Storybook 10. Renaming it to
/manager and separateing CJS and ESM exports in package.json helped
loading it in the managed mode again.

BREAKING CHANGE: If you refeirred directly to `@prantlf/storycap/register`,
remove the `/register` path. Use just `@prantlf/storycap` in the `addons`
property in `main.js`. It was the usual and documented way to register
this addons, so you probably did it anyway and this change won't be
breaking for you.